fix: prevent file traversal using helm file values param and application details api
