+++ reposerver/repository/repository_test.go
@@ -909,9 +909,7 @@ func TestGenerateHelmWithValuesDirectoryTraversalOutsideRepo(t *testing.T) {
 	})
 }
 
-// The requested file parameter (`/tmp/external-secret.txt`) is outside the app path
-// (`./util/helm/testdata/redis`), and outside the repo directory. It is used as a means
-// of providing direct content to a helm chart via a specific key.
+// File parameter should not allow traversal outside of the repository root
 func TestGenerateHelmWithAbsoluteFileParameter(t *testing.T) {
 	service := newService("../..")
 
@@ -933,16 +931,14 @@ func TestGenerateHelmWithAbsoluteFileParameter(t *testing.T) {
 			Helm: &argoappv1.ApplicationSourceHelm{
 				ValueFiles: []string{"values-production.yaml"},
 				Values:     `cluster: {slaveCount: 2}`,
-				FileParameters: []argoappv1.HelmFileParameter{
-					argoappv1.HelmFileParameter{
-						Name: "passwordContent",
-						Path: externalSecretPath,
-					},
-				},
+				FileParameters: []argoappv1.HelmFileParameter{{
+					Name: "passwordContent",
+					Path: externalSecretPath,
+				}},
 			},
 		},
 	})
-	assert.NoError(t, err)
+	assert.Error(t, err)
 }
 
 // The requested file parameter (`../external/external-secret.txt`) is outside the app path
@@ -1733,175 +1729,4 @@ func TestResolveRevisionNegativeScenarios(t *testing.T) {
 	assert.NotNil(t, err)
 	assert.Equal(t, expectedResolveRevisionResponse, resolveRevisionResponse)
 
-}
-
-func Test_resolveSymlinkRecursive(t *testing.T) {
-	testsDir, err := filepath.Abs("./testdata/symlinks")
-	if err != nil {
-		panic(err)
-	}
-	t.Run("Resolve non-symlink", func(t *testing.T) {
-		r, err := resolveSymbolicLinkRecursive(testsDir+"/foo", 2)
-		assert.NoError(t, err)
-		assert.Equal(t, testsDir+"/foo", r)
-	})
-	t.Run("Successfully resolve symlink", func(t *testing.T) {
-		r, err := resolveSymbolicLinkRecursive(testsDir+"/bar", 2)
-		assert.NoError(t, err)
-		assert.Equal(t, testsDir+"/foo", r)
-	})
-	t.Run("Do not allow symlink at all", func(t *testing.T) {
-		r, err := resolveSymbolicLinkRecursive(testsDir+"/bar", 0)
-		assert.Error(t, err)
-		assert.Equal(t, "", r)
-	})
-	t.Run("Error because too nested symlink", func(t *testing.T) {
-		r, err := resolveSymbolicLinkRecursive(testsDir+"/bam", 2)
-		assert.Error(t, err)
-		assert.Equal(t, "", r)
-	})
-	t.Run("No such file or directory", func(t *testing.T) {
-		r, err := resolveSymbolicLinkRecursive(testsDir+"/foobar", 2)
-		assert.NoError(t, err)
-		assert.Equal(t, testsDir+"/foobar", r)
-	})
-}
-
-func Test_isURLSchemeAllowed(t *testing.T) {
-	type testdata struct {
-		name     string
-		scheme   string
-		allowed  []string
-		expected bool
-	}
-	var tts []testdata = []testdata{
-		{
-			name:     "Allowed scheme matches",
-			scheme:   "http",
-			allowed:  []string{"http", "https"},
-			expected: true,
-		},
-		{
-			name:     "Allowed scheme matches only partially",
-			scheme:   "http",
-			allowed:  []string{"https"},
-			expected: false,
-		},
-		{
-			name:     "Scheme is not allowed",
-			scheme:   "file",
-			allowed:  []string{"http", "https"},
-			expected: false,
-		},
-		{
-			name:     "Empty scheme with valid allowances is forbidden",
-			scheme:   "",
-			allowed:  []string{"http", "https"},
-			expected: false,
-		},
-		{
-			name:     "Empty scheme with empty allowances is forbidden",
-			scheme:   "",
-			allowed:  []string{},
-			expected: false,
-		},
-		{
-			name:     "Some scheme with empty allowances is forbidden",
-			scheme:   "file",
-			allowed:  []string{},
-			expected: false,
-		},
-	}
-	for _, tt := range tts {
-		t.Run(tt.name, func(t *testing.T) {
-			r := isURLSchemeAllowed(tt.scheme, tt.allowed)
-			assert.Equal(t, tt.expected, r)
-		})
-	}
-}
-
-var allowedHelmRemoteProtocols = []string{"http", "https"}
-
-func Test_resolveHelmValueFilePath(t *testing.T) {
-	t.Run("Resolve normal relative path into absolute path", func(t *testing.T) {
-		p, remote, err := resolveHelmValueFilePath("/foo/bar", "/foo", "baz/bim.yaml", allowedHelmRemoteProtocols)
-		assert.NoError(t, err)
-		assert.False(t, remote)
-		assert.Equal(t, "/foo/bar/baz/bim.yaml", p)
-	})
-	t.Run("Resolve normal relative path into absolute path", func(t *testing.T) {
-		p, remote, err := resolveHelmValueFilePath("/foo/bar", "/foo", "baz/../../bim.yaml", allowedHelmRemoteProtocols)
-		assert.NoError(t, err)
-		assert.False(t, remote)
-		assert.Equal(t, "/foo/bim.yaml", p)
-	})
-	t.Run("Error on path resolving outside repository root", func(t *testing.T) {
-		p, remote, err := resolveHelmValueFilePath("/foo/bar", "/foo", "baz/../../../bim.yaml", allowedHelmRemoteProtocols)
-		assert.Error(t, err)
-		assert.Contains(t, err.Error(), "outside repository root")
-		assert.False(t, remote)
-		assert.Equal(t, "", p)
-	})
-	t.Run("Return verbatim URL", func(t *testing.T) {
-		url := "https://some.where/foo,yaml"
-		p, remote, err := resolveHelmValueFilePath("/foo/bar", "/foo", url, allowedHelmRemoteProtocols)
-		assert.NoError(t, err)
-		assert.True(t, remote)
-		assert.Equal(t, url, p)
-	})
-	t.Run("URL scheme not allowed", func(t *testing.T) {
-		url := "file:///some.where/foo,yaml"
-		p, remote, err := resolveHelmValueFilePath("/foo/bar", "/foo", url, allowedHelmRemoteProtocols)
-		assert.Error(t, err)
-		assert.False(t, remote)
-		assert.Equal(t, "", p)
-	})
-	t.Run("Implicit URL by absolute path", func(t *testing.T) {
-		p, remote, err := resolveHelmValueFilePath("/foo/bar", "/foo", "/baz.yaml", allowedHelmRemoteProtocols)
-		assert.NoError(t, err)
-		assert.False(t, remote)
-		assert.Equal(t, "/foo/baz.yaml", p)
-	})
-	t.Run("Relative app path", func(t *testing.T) {
-		p, remote, err := resolveHelmValueFilePath(".", "/foo", "/baz.yaml", allowedHelmRemoteProtocols)
-		assert.NoError(t, err)
-		assert.False(t, remote)
-		assert.Equal(t, "/foo/baz.yaml", p)
-	})
-	t.Run("Relative repo path", func(t *testing.T) {
-		c, err := os.Getwd()
-		require.NoError(t, err)
-		p, remote, err := resolveHelmValueFilePath(".", ".", "baz.yaml", allowedHelmRemoteProtocols)
-		assert.NoError(t, err)
-		assert.False(t, remote)
-		assert.Equal(t, c+"/baz.yaml", p)
-	})
-	t.Run("Overlapping root prefix without trailing slash", func(t *testing.T) {
-		p, remote, err := resolveHelmValueFilePath(".", "/foo", "../foo2/baz.yaml", allowedHelmRemoteProtocols)
-		assert.Error(t, err)
-		assert.Contains(t, err.Error(), "outside repository root")
-		assert.False(t, remote)
-		assert.Equal(t, "", p)
-	})
-	t.Run("Overlapping root prefix with trailing slash", func(t *testing.T) {
-		p, remote, err := resolveHelmValueFilePath(".", "/foo/", "../foo2/baz.yaml", allowedHelmRemoteProtocols)
-		assert.Error(t, err)
-		assert.Contains(t, err.Error(), "outside repository root")
-		assert.False(t, remote)
-		assert.Equal(t, "", p)
-	})
-	t.Run("Garbage input as values file", func(t *testing.T) {
-		p, remote, err := resolveHelmValueFilePath(".", "/foo/", "kfdj\\ks&&&321209.,---e32908923%$§!\"", allowedHelmRemoteProtocols)
-		assert.Error(t, err)
-		assert.Contains(t, err.Error(), "outside repository root")
-		assert.False(t, remote)
-		assert.Equal(t, "", p)
-	})
-	t.Run("NUL-byte path input as values file", func(t *testing.T) {
-		p, remote, err := resolveHelmValueFilePath(".", "/foo/", "\000", allowedHelmRemoteProtocols)
-		assert.Error(t, err)
-		assert.Contains(t, err.Error(), "outside repository root")
-		assert.False(t, remote)
-		assert.Equal(t, "", p)
-	})
 }
