+++ util/helm/testdata/symlinks/bar
