+++ util/helm/testdata/symlinks/foo
